Re: BFD_INIT_MAGIC
I should know better than to introduce the first use of size_t in
bfd.h.

	PR 23534
	* init.c (bfd_init): Return an unsigned int.
	bfd-in2.h: Regenerate.

+++ bfd/bfd-in2.h
@@ -1080,7 +1080,7 @@ extern void elf32_csky_next_input_section
 extern int elf32_csky_setup_section_lists
   (bfd *, struct bfd_link_info *);
 /* Extracted from init.c.  */
-size_t bfd_init (void);
+unsigned int bfd_init (void);
 
 
 /* Value returned by bfd_init.  */
